Bot presence condition add to bot killer script.

+++ kill_bot.sh
@@ -1,2 +1,11 @@
+#!/bin/bash
+
 bot_pid=$(ps aux | grep 'python bot.py$' | cut -d' ' -f2)
-kill -9 $bot_pid
+
+if [ -z $bot_pid ]
+then
+	echo "Bot process is not found!"	
+else
+	kill -9 $bot_pid
+fi
+
